removed timepoint function from dsp_plot.py and create stand along script, timing_dsp.py
